Check if `infoFile` exists before importing
When `drvs` is set and `mvnix-update` is being run to generate the
initial `infoFile`, then `drvsInfo` will be non-empty. Because
`drvsInfo` is now non-empty, `transRemotes drvsInfo` (in `mkRepo`) will
now be evaluated (when `drvsInfo` is empty, the `foldl'` is a no-op).
`transRemotes` references `info.remotes`, and `info` is defined as
importing the `infoFile` (which we're trying to generate!).

This commit fixes this by checking if the path to `infoFile` exists
before attempting to import it.

Signed-off-by: David Wood <david@davidtw.co>